hotfix on the parseFloat value

+++ frontend/src/pages/UpdateItem/UpdateItem.jsx
@@ -28,7 +28,7 @@ const UpdateItem = () => {
             id: parseInt(id),
             name,
             description,
-            value: parseInt(value),
+            value: parseFloat(value),
         })
 
         navigate("/items")
